requirements.txt
pandas
opencv-python
pillow
torch
torchvision
scikit-image
pycryptodome
tritonclient[all]
onnx
onnxruntime
tensorrt
